finished reducer generator
