+++ generators/templates/components/stories.hbs
@@ -10,7 +10,7 @@ export default {
 
 const Template: ComponentStory<typeof {{ properCase name }}> = (args) => (
   <>
-      <{{ properCase name }} {...args} />
+    <{{ properCase name }} {...args} />
   </>
 );
 export const Default = Template.bind({});
