+++ generators/templates/components/styles.hbs
@@ -2,4 +2,4 @@
 
 .{{camelCase name}} { 
 
-  }
+}
