+++ src/store/rootReducer.ts
@@ -1,9 +1,11 @@
 import ui from './ui/reducer';
 import modals from './modals/reducer';
 import user from './user/reducer';
+/* PLOP_INJECT_IMPORT_REDUCER */
 
 export default {
   ui,
   user,
   modals,
+  /* PLOP_INJECT_PLACE_REDUCER */
 };
